Adds Boundary activity structure
