+++ app/src/org/commcare/gis/WalkingBoundaryActivity.kt
@@ -1,0 +1,9 @@
+package org.commcare.gis
+
+class WalkingBoundaryActivity : BaseKujakuActivity() {
+
+    override fun onMapLoaded() {
+        TODO("Not yet implemented")
+    }
+
+}
